refactor: review nits — timing constants, file consolidation, toast constant
- integration_timing.go: Datar flush wait derived from flush interval + margin
- DUPLICATE_SUCCESS_TOAST_DURATION_MS; Flag.vue documents route watcher mount
- AppendTagValueToFlag in flag_template.go; drop flag_clone*
- commitFlagMutation in crud.go; drop crud_snapshot.go
- validatePutVariantForDistributions: document tx for atomic distribution key sync
- plan/overview path updates
